Added a wraparound board class
Further subclassing of Board, with an abstract grid-based board and a
wraparound board class.  Most of the work done by the constructor is
moved to methods which can be separately overridden.  The constructor
that reads a Board from a file is disabled until it can be cleaned up
and rewritten to use the new subclass structure.